+++ Engine/src/platforms/windows/window.h
@@ -9,8 +9,8 @@ class WindowsWindow : public Window {
     GraphicContext *graphic_ctx;
     GLFWwindow *window;
     bool vsync = false;
-    double last_time = 0.0;
-    double last_frame_time = 0.0;
+    float last_time = 0.0;
+    float last_frame_time = 0.0;
 
 public:
     WindowsWindow(const char *title, unsigned int width, unsigned int height);
@@ -20,7 +20,7 @@ public:
     unsigned get_height() const override;
     bool is_vsync() const override;
     void set_vsync(bool vsync) override;
-    double get_last_frame_time() const override;
+    float get_last_frame_time() const override;
 
     GLFWwindow *get_native_window() const override {
         return this->window;
